+++ libiti/srcs/string2number.c
@@ -6,7 +6,7 @@
 /*   By: mel-omar <mel-omar@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/05 16:10:24 by mel-omar          #+#    #+#             */
-/*   Updated: 2021/04/13 13:30:25 by mel-omar         ###   ########.fr       */
+/*   Updated: 2021/04/22 16:06:43 by mel-omar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	isdigits(char c)
 	return (0);
 }
 
-int			string2number(const char *string)
+long	string2number(const char *string)
 {
 	long int	number;
 	int			sign;
@@ -28,7 +28,7 @@ int			string2number(const char *string)
 		return (0);
 	else if (*string == '-')
 	{
-		sign  = -1;
+		sign = -1;
 		string++;
 	}
 	else if (*string == '+')
